Refactor: Reorganize C++ source code into logical directory structure
Reorganized the entire codebase from a flat src/ directory into a
hierarchical structure that separates concerns and improves maintainability.

Changes:
- Created lib/ directory for reusable, project-independent libraries:
  - lib/logger/ - Thread-safe singleton logger
  - lib/colormod/ - Terminal color utilities
  - lib/distance/ - Levenshtein distance algorithm
  - lib/property/ - Generic property template

- Organized src/ into logical subdirectories:
  - src/core/ - Game domain objects (game, escenario, objeto, jugador, etc.)
  - src/patterns/factory/ - Factory pattern implementation
  - src/patterns/command/ - Command pattern implementation
  - src/events/ - Event system (ievent, event, eventsqueue, etc.)
  - src/manager/ - Application management (manager, parametros)
  - src/tests/ - All unit tests (*_test.cpp files)

- Updated all #include paths to reflect new structure:
  - lib/ files use local includes
  - src/ files use relative paths (../core/, ../../events/, etc.)
  - Added lib/ references where needed (lib/logger/logger.h, etc.)

- Rewrote Makefile for new structure:
  - Added recursive source discovery with find command
  - Updated include paths: -I. -Isrc -Ilib -I3rdParty/tinyxml
  - Improved clean targets to work with new directory structure
  - Added individual test build targets for convenience

Benefits:
- Clear separation between reusable libraries and game-specific code
- Improved code discoverability and navigation
- Easier to understand dependencies and architecture
- Design patterns explicitly organized and documented
- Tests isolated from production code
- Build system automatically finds all sources

All code compiles successfully. No functional changes.
