+++ lib/colormod/colormod.cpp
