+++ src/core/entidad.cpp
