+++ src/core/entidad.h
