+++ src/core/escenario.cpp
