+++ src/core/escenario.h
