+++ src/core/inventario.cpp
