+++ src/core/inventario.h
