+++ src/core/objeto.cpp
