+++ src/events/activador.cpp
