+++ src/events/conceal.cpp
