+++ src/manager/parametros.cpp
