+++ src/manager/parametros.h
