+++ src/patterns/command/ICommand.cpp
