+++ src/patterns/command/ICommand.h
