+++ src/patterns/factory/abstract_creator.cpp
