+++ src/patterns/factory/factory.cpp
