+++ src/tests/conceal_test.cpp
@@ -1,4 +1,4 @@
-#include "conceal.h"
+#include "../events/conceal.h"
 
 
 int main(){
