+++ src/tests/property_test.cpp
@@ -1,4 +1,4 @@
-#include "property.h"
+#include "lib/property/property.h"
 
 #include <iostream>
 #include <string>
